+++ src/kard/users/rewards/types/__init__.py
@@ -49,6 +49,8 @@ if typing.TYPE_CHECKING:
     from .operation_time import OperationTime
     from .placement_batch_attributes import PlacementBatchAttributes
     from .placement_batch_data import PlacementBatchData
+    from .placement_content_data import PlacementContentData
+    from .placement_content_response import PlacementContentResponse
     from .progress_bar import ProgressBar
     from .progress_bar_label_pair import ProgressBarLabelPair
     from .progress_bar_labels import ProgressBarLabels
@@ -106,6 +108,8 @@ _dynamic_imports: typing.Dict[str, str] = {
     "OperationTime": ".operation_time",
     "PlacementBatchAttributes": ".placement_batch_attributes",
     "PlacementBatchData": ".placement_batch_data",
+    "PlacementContentData": ".placement_content_data",
+    "PlacementContentResponse": ".placement_content_response",
     "ProgressBar": ".progress_bar",
     "ProgressBarLabelPair": ".progress_bar_label_pair",
     "ProgressBarLabels": ".progress_bar_labels",
@@ -187,6 +191,8 @@ __all__ = [
     "OperationTime",
     "PlacementBatchAttributes",
     "PlacementBatchData",
+    "PlacementContentData",
+    "PlacementContentResponse",
     "ProgressBar",
     "ProgressBarLabelPair",
     "ProgressBarLabels",